+++ src/main.rs
@@ -1,5 +1,5 @@
 mod build;
-mod metadata;
+mod models;
 mod ogp;
 mod serve;
 mod templates;
